Added sync transactions to the server.
The client sends the server a list of all newly created and updated
entities since the last time it synced. The server sends back all
newly created and updated entities since the client synced last time.
The client receives the array and determines whether each entity is
new or an updated version of one it already knows about and either
creates a new managed object in CD or updates an existing one.