Marking entities as conflicted when server detects conflicted state
If the lastModifiedTime for an entity the client is attempting
to sync to the server is earlier than the LMT  the server has for
that entity, that means the clients version is out of date. The
server rejects the sync for that object and send back the server
version in the 'conflictedEntities' array. The client is responsible
for resolving the conflict and resyncing the object.