Implementing a diffing of 2 different SyncObject subclasses
When there is a conflict discovered we need to find out what the
actual diffs are between the server version and the client version
so they can be presented to the user and resolved. The diff will
be presented to the client code of SyncStorageManger as a dict
with all the differing attributes as keys. The values will contain
the server and the client versions.