Conflict resolution implemented on the client
When the server sends back a conflict to the client, the client
code is presented with the diffs between the clients version of the
resource and the servesr version. Once the diffs are resolved by
the user, resolve: is called. resolve: sets the values to the
resolved values in the database, removes the conflict and sets the
newly created object's state to 'needs sync' so that the resolution
will be propegated to the server.